migrates out read-private/make-private from js api
